Add files via upload
A bus in a city starts and ends at the same point. The bus stops at multiple points in the city. If you have to travel from starting point to a point in between. Let's say 5th stop, you have to take the same decided route and cover all the points in between.
For every 1km charge rs 5 and the minimum fare is rs 20 means if a person is travelling between two points and the bus fare is calculated to rs 7.5, the conductor should collect rs 20 but if the bus fare calculates more than rs 20 the calculated amount should be calculated.
round off the decimal fare values and give an integer value in the output.

Write a code in Java with user input for the number of points and distance between each point to find the bus fare for the distance travelled between any two points.